+++ packages/jest-fake-timers/src/legacyFakeTimers.ts
@@ -134,6 +134,10 @@ export default class FakeTimers<TimerRef = unknown> {
     this._timers = new Map();
   }
 
+  now(): number {
+    return this._now;
+  }
+
   runAllTicks(): void {
     this._checkFakeTimers();
     // Only run a generous number of ticks and then bail.
@@ -200,13 +204,15 @@ export default class FakeTimers<TimerRef = unknown> {
     // This is just to help avoid recursive loops
     let i;
     for (i = 0; i < this._maxLoops; i++) {
-      const nextTimerHandle = this._getNextTimerHandle();
+      const nextTimerHandleAndExpiry = this._getNextTimerHandleAndExpiry();
 
       // If there are no more timer handles, stop!
-      if (nextTimerHandle === null) {
+      if (nextTimerHandleAndExpiry === null) {
         break;
       }
 
+      const [nextTimerHandle, expiry] = nextTimerHandleAndExpiry;
+      this._now = expiry;
       this._runTimerHandle(nextTimerHandle);
 
       // Some of the immediate calls could be enqueued
@@ -239,7 +245,10 @@ export default class FakeTimers<TimerRef = unknown> {
 
     timerEntries
       .sort(([, left], [, right]) => left.expiry - right.expiry)
-      .forEach(([timerHandle]) => this._runTimerHandle(timerHandle));
+      .forEach(([timerHandle, timer]) => {
+        this._now = timer.expiry;
+        this._runTimerHandle(timerHandle);
+      });
   }
 
   advanceTimersToNextTimer(steps = 1): void {
@@ -265,21 +274,16 @@ export default class FakeTimers<TimerRef = unknown> {
     // This is just to help avoid recursive loops
     let i;
     for (i = 0; i < this._maxLoops; i++) {
-      const timerHandle = this._getNextTimerHandle();
+      const timerHandleAndExpiry = this._getNextTimerHandleAndExpiry();
 
       // If there are no more timer handles, stop!
-      if (timerHandle === null) {
+      if (timerHandleAndExpiry === null) {
         break;
       }
-      const timerValue = this._timers.get(timerHandle);
-      if (timerValue === undefined) {
-        break;
-      }
-      const nextTimerExpiry = timerValue.expiry;
+      const [timerHandle, nextTimerExpiry] = timerHandleAndExpiry;
+
       if (this._now + msToRun < nextTimerExpiry) {
-        // There are no timers between now and the target we're running to, so
-        // adjust our time cursor and quit
-        this._now += msToRun;
+        // There are no timers between now and the target we're running to
         break;
       } else {
         msToRun -= nextTimerExpiry - this._now;
@@ -287,6 +291,9 @@ export default class FakeTimers<TimerRef = unknown> {
         this._runTimerHandle(timerHandle);
       }
     }
+
+    // Advance the clock by whatever time we still have left to run
+    this._now += msToRun;
 
     if (i === this._maxLoops) {
       throw new Error(
@@ -557,7 +564,7 @@ export default class FakeTimers<TimerRef = unknown> {
     return this._timerConfig.idToRef(uuid);
   }
 
-  private _getNextTimerHandle() {
+  private _getNextTimerHandleAndExpiry(): [string, number] | null {
     let nextTimerHandle = null;
     let soonestTime = MS_IN_A_YEAR;
 
@@ -568,13 +575,19 @@ export default class FakeTimers<TimerRef = unknown> {
       }
     });
 
-    return nextTimerHandle;
+    if (nextTimerHandle === null) {
+      return null;
+    }
+
+    return [nextTimerHandle, soonestTime];
   }
 
   private _runTimerHandle(timerHandle: TimerID) {
     const timer = this._timers.get(timerHandle);
 
     if (!timer) {
+      // Timer has been cleared - we'll hit this when a timer is cleared within
+      // another timer in runOnlyPendingTimers
       return;
     }