+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ### Features
 
+- `[@jest/fake-timers]` Add `jest.now()` to return the current fake clock time ([#13244](https://github.com/facebook/jest/pull/13244))
+
 ### Fixes
 
 ### Chore & Maintenance